requirements.txt
# YOLOv8 训练依赖
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# 数据处理
numpy>=1.20.0
Pillow>=9.0.0
opencv-python>=4.5.0

# 配置解析
PyYAML>=5.4.0

# ONNX 推理 (可选)
onnxruntime>=1.15.0
onnx>=1.14.0

# 可视化
matplotlib>=3.3.0
tqdm>=4.60.0

# 其他
requests>=2.25.0
psutil>=5.8.0
